+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2022-05-16
+
+### Added
+
+- Associative array tree to materialized path convertion
+- Associative array tree to materialized path convertion test
+
 ## [0.7.0] - 2022-05-13
 
 ### Added
 
 - Associative array tree to ajacency list convertion
-- Associative array tree to ajacency list convertion tests
+- Associative array tree to ajacency list convertion test
 - `$noParentValue` parameter for NS to AL convertion
 
 ## [0.6.0] - 2022-04-22